pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fusionUAV"
version = "0.1.0"
description = "AirSim UAV sensor fusion runtime with ESKF estimation, control, plotting, and logging."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=2.0",
    "airsim",
    "PyYAML>=6.0",
    "PyQt5>=5.15",
    "keyboard>=0.13.5",
    "msgpack-rpc-python>=0.4.1",
    "pyqtgraph>=0.13",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
